pyproject.toml
[project]
name = "graph-benchmarks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.8",
    "pyyaml>=6.0.3",
    "requests[socks]>=2.33.1",
    "scipy>=1.17.1",
    "typer>=0.24.1",
]
